renderPageBox方法算出每页的宽高
