+++ src/utils/ofd_render.js
@@ -18,3 +18,67 @@
  *
  */
 
+import {converterDpi, setPageScal} from "@/utils/ofd_util";
+
+export const renderPageBox = function (screenWidth, pages, document) {
+    let pageBoxs = [];
+    for (const page of pages) {
+        const area = page[Object.keys(page)[0]]['json']['ofd:Area'];
+        let box;
+        if (area) {
+            const physicalBox = area['ofd:PhysicalBox']
+            if (physicalBox) {
+                box = (physicalBox);
+            } else {
+                const applicationBox = area['ofd:ApplicationBox']
+                if (applicationBox) {
+                    box = (applicationBox);
+                } else {
+                    const contentBox = area['ofd:ContentBox']
+                    if (contentBox) {
+                        box = (contentBox);
+                    }
+                }
+            }
+        } else {
+            let documentArea = document['ofd:CommonData']['ofd:PageArea']
+            const physicalBox = documentArea['ofd:PhysicalBox']
+            if (physicalBox) {
+                box = (physicalBox);
+            } else {
+                const applicationBox = documentArea['ofd:ApplicationBox']
+                if (applicationBox) {
+                    box = (applicationBox);
+                } else {
+                    const contentBox = documentArea['ofd:ContentBox']
+                    if (contentBox) {
+                        box = (contentBox);
+                    }
+                }
+            }
+        }
+        box = parsePageBox(screenWidth, box);
+        let boxObj = {};
+        boxObj['id'] = Object.keys(page)[0];
+        boxObj['box'] = box;
+        pageBoxs.push(boxObj);
+    }
+    return pageBoxs;
+}
+
+const parsePageBox = function (screenWidth, obj) {
+    if (obj) {
+        let array = obj.split(' ');
+        let width = converterDpi(parseFloat(array[2]));
+        if (width > screenWidth) {
+            const scale = (screenWidth - 5) / parseFloat(array[2]);
+            setPageScal(scale > 0 ? scale : 1);
+        }
+        return {
+            x: converterDpi(parseFloat(array[0])), y: converterDpi(parseFloat(array[1])),
+            w: converterDpi(parseFloat(array[2])), h: converterDpi(parseFloat(array[3]))
+        };
+    } else {
+        return null;
+    }
+}
